+++ packages/media/media-test-helpers/src/mocks/media-mock.ts
@@ -23,6 +23,8 @@ export class MediaMock {
 
   disable() {
     // TODO: add teardown logic to kakapo server
+    // tslint:disable:no-console
+    console.warn('Disabling logic is not implemented in MediaMock');
   }
 }